refactor: MainActivity app 모듈 -> feature:main 모듈로 이동시킴
