+++ feature/main/src/main/java/com/kusitms/connectdog/feature/main/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.kusitms.connectdog
+package com.kusitms.connectdog.feature.main
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -10,13 +10,12 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.kusitms.connectdog.ui.theme.ConnectDogTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            ConnectDogTheme {
+            com.kusitms.connectdog.core.designsystem.theme.ConnectDogTheme {
                 // A surface container using the 'background' color from the theme
                 Surface(
                     modifier = Modifier.fillMaxSize(),
@@ -40,7 +39,7 @@ fun Greeting(name: String, modifier: Modifier = Modifier) {
 @Preview(showBackground = true)
 @Composable
 fun GreetingPreview() {
-    ConnectDogTheme {
+    com.kusitms.connectdog.core.designsystem.theme.ConnectDogTheme {
         Greeting("Android")
     }
 }
